requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu117

torch==2.0.1+cu117
torchvision==0.15.2+cu117

anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cmake==4.2.1
coloredlogs==15.0.1
einops==0.8.2
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
kornia==0.8.2
kornia_rs==0.1.10
lit==18.1.8
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4

nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91

onnxruntime-gpu==1.16.3
opencv-python==4.8.1.78
packaging==26.0
pillow==12.1.0
protobuf==6.33.5
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
shellingham==1.5.4
sympy==1.14.0
timm==1.0.24
tokenizers==0.19.1
tqdm==4.67.3
transformers==4.40.2
triton==2.0.0
typer-slim==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
